Adds setting and flag to keep generated LaTeX book template

+++ compiler/book.go
@@ -70,7 +70,9 @@ func MakeBook(path string, opts *settings.Settings) {
 			"error":   err,
 		}).Error("failed to clean songbook latex files")
 	}
-	err = os.Remove(texPath)
+	if !opts.KeepBookTemplate {
+		err = os.Remove(texPath)
+	}
 	helpers.Check(err, "could not remove songbook latex template")
 }
